+++ CHANGELOG.rst
@@ -14,13 +14,13 @@ Improvements
 - Add env variable ``BLUEETL_SUBPROCESS_LOGGING_LEVEL`` to set a logging level in subprocesses.
 - Improve log of execution times.
 - Improve performances of ``etl.q`` when only a single condition is specified.
+- Lock the cache used by the Analyzer instance [NSETM-1971]
 
 Bug Fixes
 ~~~~~~~~~
 - Reset the index in the simulations dataframe after applying filters.
 - Ensure that RangeIndex is converted to Int64Index in MultiIndexes with Pandas 1.5.0,
   see https://issues.apache.org/jira/browse/ARROW-17806.
-
 
 Version 0.1.0
 -------------